refactor: move old service to use-cases
